refactor(page): consolidate components, data, and types into ServiceHistory
